Move utilities to 'utils' package and fix visibility on some functions
