+++ library/src/main/java/io/github/pelmenstar1/rangecalendar/utils/ColorHelper.kt
@@ -1,15 +1,15 @@
-package io.github.pelmenstar1.rangecalendar
+package io.github.pelmenstar1.rangecalendar.utils
 
 import android.graphics.Color
 import androidx.annotation.ColorInt
 import kotlin.math.roundToInt
 
 @ColorInt
-fun Int.withAlpha(alpha: Int): Int {
+internal fun Int.withAlpha(alpha: Int): Int {
     return this and 0x00FFFFFF or (alpha shl 24)
 }
 
-fun colorLerp(@ColorInt start: Int, @ColorInt end: Int, fraction: Float): Int {
+internal fun colorLerp(@ColorInt start: Int, @ColorInt end: Int, fraction: Float): Int {
     return Color.argb(
         lerp(Color.alpha(start), Color.alpha(end), fraction),
         lerp(Color.red(start), Color.red(end), fraction),
@@ -19,7 +19,7 @@ fun colorLerp(@ColorInt start: Int, @ColorInt end: Int, fraction: Float): Int {
 }
 
 @ColorInt
-fun Int.darkerColor(factor: Float): Int {
+internal fun Int.darkerColor(factor: Float): Int {
     val invFactor = 1f - factor
 
     return Color.argb(
